% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbooks.R
\name{get_link}
\alias{get_link}
\title{Get a book link.}
\usage{
get_link(name, browse = TRUE)
}
\arguments{
\item{name}{URL name}

\item{browse}{Open browser}
}
\value{
Prints URL
}
\description{
The function copies the URL and
opens the link in the browser.
}
